Updated site to use library controller and organized file better
