+++ composer.json
@@ -10,8 +10,9 @@
     ],
     "require": {
         "php": ">=5.6.0",
-        "slim/slim": "^3.5",
-        "dschoenbauer/view": "^1.0"
+        "slim/slim": "^3.6",
+        "dschoenbauer/view": "^1.0",
+        "dschoenbauer/slim-controller": "^1.0"
     },
     "autoload": {
         "psr-4": {
