+++ src/Klik/Pages/Faq.php
@@ -1,13 +1,7 @@
-<?php
+<?php namespace DSchoenbauer\Klik\Pages;
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-namespace DSchoenbauer\Klik;
-
+use DSchoenbauer\Controller\AbstractController;
+use DSchoenbauer\Klik\Component\Layout;
 use DSchoenbauer\View\TemplatedView;
 use Slim\Http\Request;
 use Slim\Http\Response;
