--- src/Klik/VisiteeInterface.php
@@ -1,10 +1,0 @@
-<?php namespace DSchoenbauer\Klik;
-
-/**
- * Description of VisiteeInterface
- *
- * @author David
- */
-interface VisiteeInterface {
-    public function accept(VisitorInterface $visitor);
-}
