--- src/Klik/VisitorInterface.php
@@ -1,11 +1,0 @@
-<?php namespace DSchoenbauer\Klik;
-
-use Slim\App;
-
-/**
- *
- * @author David
- */
-interface VisitorInterface {
-    public function visitApp(App $app);
-}
